feat: Add distillation pipeline and update training scripts
- Introduced a new distillation pipeline with structured directory organization for experiments.
- Added JSON summary files for distillation runs and student model checkpoints.
- Created comprehensive command guides for Chronos and TimeLLM training and inference.
- Implemented knowledge distillation logic in main.py, integrating with existing LLM frameworks.
- Updated student model and TimeLLM classes to remove unnecessary dtype conversions.
- Removed outdated distillation scripts and refactored config generation for clarity and efficiency.
- Enhanced logging and error handling throughout the pipeline for better traceability.

+++ models/layers/student_model.py
@@ -128,7 +128,7 @@ class Model(nn.Module):
         ).permute(1, 0)
 
         x_enc = x_enc.permute(0, 2, 1).contiguous()
-        enc_out, n_vars = self.patch_embedding(x_enc.to(torch.bfloat16))
+        enc_out, n_vars = self.patch_embedding(x_enc)
         enc_out = self.reprogramming_layer(
             enc_out, source_embeddings, source_embeddings
         )
